requirements.txt
fastapi
uvicorn
SQLAlchemy
alembic
psycopg2-binary
python-jose
passlib
bcrypt
pytest
httpx
locust
